tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.rect.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.rect.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.rect.negative</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.beginPath();
  ctx.fillStyle = '#0f0';
  ctx.rect(0, 0, 50, 25);
  ctx.rect(100, 0, -50, 25);
  ctx.rect(0, 50, 50, -25);
  ctx.rect(100, 50, -50, -25);
  ctx.fill();
  _assertPixel(canvas, 25,12, 0,255,0,255);
  _assertPixel(canvas, 75,12, 0,255,0,255);
  _assertPixel(canvas, 25,37, 0,255,0,255);
  _assertPixel(canvas, 75,37, 0,255,0,255);
  t.done();

});
</script>
